+++ src/component/AppLayout/Main.jsx
@@ -110,8 +110,11 @@ export class Main extends AbstractComponent {
                 onClick={ this.onExtClick }>commands, assertions and references</a> { "" }
               to ensure that a targeted component of the test application acts as intended.
               Tests cases are organized into logical { "" } <a href="https://docs.puppetry.app/group"
-                onClick={ this.onExtClick }>groups</a>, representing test contexts.
-
+                onClick={ this.onExtClick }>groups</a>, representing test contexts. During the test
+                execution <a
+                onClick={ this.onExtClick }
+                href="https://pptr.dev/#?product=Puppeteer&version=v2.0.0&show=api-class-browser">
+                a new browser session</a> is created for every group.
               </p>
 
               <p>