+++ src/component/AppLayout/Main/GroupTable/TestTable/CommandRowLabel.jsx
@@ -4,7 +4,7 @@ import PropTypes from "prop-types";
 import { Icon } from "antd";
 import Tooltip from "component/Global/Tooltip";
 import { getSchema } from "component/Schema/schema";
-import { truncate } from "service/utils";
+import { truncate, extendToGherkin } from "service/utils";
 import { connect } from "react-redux";
 
 // Mapping state to the props
@@ -63,7 +63,7 @@ export class CommandRowLabel extends React.Component {
          { title }
        </span>
        { record.comment && <i className="is-optional comment-label">
-         <Icon type="message" title="Comment" />{ " " } { record.comment }</i> }
+       <Icon type="message" title="Comment" /><span>{ record.comment }</span></i> }
      </div> );
    }
 
@@ -100,17 +100,24 @@ export class CommandRowLabel extends React.Component {
          pos="up" />
 
        { testStyle === "gherkin" && <div className="gherkin">
-         { CommandRowLabel.highlightText( schema.toGherkin( record ) ) }
+         { CommandRowLabel.highlightText( extendToGherkin( record, schema.toGherkin( record ) ) ) }
        </div> }
        { testStyle !== "gherkin" && <React.Fragment>
-         <Icon
+           { record.waitForTarget && <React.Fragment>
+            <Icon
+            type={ record.target === "page" ? "file" : "scan" }
+            title={ record.target === "page" ? "Page method" : `${ record.target } target method` } />
+            <span className="token--target">{ record.target }</span>.waitForTarget
+            <span className="token--param">()</span><br />
+           </React.Fragment> }
+        <Icon
            type={ record.target === "page" ? "file" : "scan" }
            title={ record.target === "page" ? "Page method" : `${ record.target } target method` } />
          <span className="token--target">{ record.target }</span>.{ record.method }
          <span className="token--param">{ CommandRowLabel.buildAddon( record ) }</span>
        </React.Fragment> }
        { record.comment && <i className="is-optional comment-label">
-         <Icon type="message" title="Comment" />{ " " } { record.comment }</i> }
+       <Icon type="message" title="Comment" /><span>{ record.comment }</span></i> }
      </div> );
    }